傳送中閃爍玩家。新增動畫控制腳本。

+++ Assets/Scripts/UI/GameScene/CardData.cs
@@ -9,7 +9,7 @@ namespace TBL.UI.GameScene
     public class CardData : MonoBehaviour
     {
         public Text nameTextUI;
-        public ushort cardID;
+        public int cardID;
         public bool isSelected = false;
 
         NetCanvas.GameScene netCanvas;
